kboot: Move seg code into libkboot
This code is really generic segment processing and should live here. It
also interfaces primarily with Linux APIs, so it's also appropriate for
this if we have a version of that runs as a FreeBSD binary using FreeBSD
system calls and APIs, though that's in the future somehow...

Sponsored by:		Netflix
