+++ stand/kboot/libkboot/Makefile
@@ -9,6 +9,7 @@ CFLAGS+=-I${.CURDIR} -I${.CURDIR}/arch/${MACHINE_ARCH}
 SRCS=	crt1.c
 SRCS+=	host_syscall.S
 SRCS+=	host_syscalls.c
+SRCS+=	seg.c
 SRCS+=	termios.c
 
 .sinclude "${.CURDIR}/arch/${MACHINE_ARCH}/Makefile.inc"
